+++ Changelog.md
@@ -17,6 +17,7 @@ Changes are in reverse chronological order; newest changes at the top.
 * Fixed Small Tanks not getting used up in crafting recipes
 * Fixed Logistics Module not rendering properly
 * Fixed Pressure Chamber crafting bug where 2 milk buckets made slime balls (correct recipe is 1 milk bucket + 4 green dye = 1 bucket + 4 slime balls)
+* Fixed Heat Frame Cooling not working on dedicated server
 
 ## 1.1.1-6 (May 5 2020)